the word "interface" is a reserved word
i changed it to interface_page. reason being, is it would error out and the bot would restart so the shop command didn't work properly also somehow this also made the profile command not work but as soon as i hot fixed it in glitch both worked. profile gave an error if no art work but other than that its fine.